redo and fix trapping efficiency TE computation along river network
Two algorithms - first walks over each node in river network and routes the runoff
generated in that node through all downstream reservoirs to calculate the
trapping efficiency for that volume of water, and then weights all the TE by
node runoff. Runoff is scaled so that total over basin matches the discharge
from the river mouth.

Other algorithm adds to this to estimate the effective discharge at each
reservoir that can be traced back to each given node. Uses modeled disharge at
each cell, and allocates that based on generated runoff in the node divided by
all runoff entering the node with a reservoir. Effective discharge through the
reservoirs is the node-allocated flow weighted by reservoir volume for each
reservoir along the network to the mouth.

Results of two algorithms are very similar, and seem much better that the bulk
estimate